% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ids_states.R
\name{extract_ids_states}
\alias{extract_ids_states}
\title{Identify patids for individuals in state j at time t}
\usage{
extract_ids_states(data.mstate, tmat, j, t)
}
\arguments{
\item{data.mstate}{Validation data in \code{msdata} format}

\item{tmat}{Transition probability matrix}

\item{j}{State j}

\item{t}{Follow up time}
}
\description{
Extract patids for individuals in state j at time t from a dataset in 'msdata'
format. Used internally in calib_blr, calib_mlr and calib_pv.
}
