% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibmsm.R
\name{metadata}
\alias{metadata}
\title{Create S3 generic for printing metadata}
\usage{
metadata(x, ...)
}
\arguments{
\item{x}{Object generated from \code{\link{calib_msm}}.}

\item{...}{Extra arguments}
}
\description{
Create S3 generic for printing metadata
}
