% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msebmtcal}
\alias{msebmtcal}
\title{European Group for Blood and Marrow Transplantation data}
\format{
\subsection{'msebmtcal'}{

A data frame in \code{msdata} format (see \code{\link[mstate]{msprep}}) with 15,512 rows and 8 columns:
\describe{
\item{id}{Patient indentifier}
\item{from}{transition from state}
\item{to}{transition to state}
\item{trans}{transition number}
\item{Tstart}{time entered state 'from'}
\item{Tstop}{time leaving state 'from'}
\item{time}{time in state 'from'}
\item{status}{event indicator, 1:transitioned to state 'to'}
}
}
}
\usage{
msebmtcal
}
\description{
The \code{\link[mstate]{ebmt4}} data converted into \code{msdata} format (see \code{\link[mstate]{msprep}}),
using the processes implemented in the \code{mstate} package.
}
\keyword{datasets}
