% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calib.R
\name{plot.calib_mlr}
\alias{plot.calib_mlr}
\title{Plots calibration scatter plots estimated using \code{\link{calc_calib_mlr}}.}
\usage{
\method{plot}{calib_mlr}(x, ..., combine = TRUE, ncol = NULL, nrow = NULL)
}
\arguments{
\item{x}{Object of class 'calib_mlr' generated from \code{\link{calc_calib_mlr}}}

\item{...}{Other}

\item{combine}{Whether to combine into one plot using ggarrange, or return as a list of individual plots}

\item{ncol}{Number of columns for combined calibration plot}

\item{nrow}{Number of rows for combined calibration plot}
}
\value{
If \code{combine = TRUE}, returns an object of classes \code{gg}, \code{ggplot}, and \code{ggarrange},
as all ggplots have been combined into one object. If \code{combine = FALSE}, returns an object of
class \code{list}, each element containing an object of class \code{gg} and \code{ggplot}.
}
\description{
Plots calibration scatter plots for the transition probabilities of a multistate model
using data estimated from \code{\link{calc_calib_mlr}}.
}
