% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msebmtcal.cmprsk}
\alias{msebmtcal.cmprsk}
\title{European Group for Blood and Marrow Transplantation data in competing risks format, for transitions out of the initial state only}
\format{
\subsection{'msebmtcal.cmprsk'}{

A data frame with 9,116 rows and 8 columns:
\describe{
\item{id}{Patient indentifier}
\item{from}{transition from state}
\item{to}{transition to state}
\item{trans}{transition number}
\item{Tstart}{time entered state 'from'}
\item{Tstop}{time leaving state 'from'}
\item{time}{time in state 'from'}
\item{status}{event indicator, 1:transitioned to state 'to'}
}
}
}
\usage{
msebmtcal.cmprsk
}
\description{
Used in vignette: Comparison-with-graphical-calibration-curves-in-competing-risks-setting.
The \code{\link[mstate]{ebmt4}} data converted into \code{msdata} format (see \code{\link[mstate]{msprep}}),
where all subsequent states are considered absorbing states. i.e. only transitions out of the initial state are considered,
meaning this data constitutes a competing risks model out of the initial state.
}
\keyword{datasets}
