% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msebmtcal.cmprsk}
\alias{msebmtcal.cmprsk}
\title{European Group for Blood and Marrow Transplantation data in competing risks format, for transitions out of the initial state only}
\format{
\subsection{'msebmtcal.cmprsk'}{

A data frame with 9,116 rows and 8 columns:
\describe{
\item{id}{Patient indentifier}
\item{from}{transition from state}
\item{to}{transition to state}
\item{trans}{transition number}
\item{Tstart}{time entered state 'from'}
\item{Tstop}{time leaving state 'from'}
\item{time}{time in state 'from'}
\item{status}{event indicator, 1:transitioned to state 'to'}
}
}
}
\usage{
msebmtcal.cmprsk
}
\description{
Used in vignette: Comparison-with-graphical-calibration-curves-in-competing-risks-setting.
This is the ebmt data in 'msdata' format, derived from the ebmt dataset using software from the mstate package.
However, only transitions out of the initial state are considered, making this data for competing risks model out of the initial state.
}
\keyword{datasets}
