% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tp.cmprsk.j0}
\alias{tp.cmprsk.j0}
\title{Predicted risks for a competing risks model out of state j = 0}
\format{
\subsection{'tp.cmprsk.j0'}{

A data frame with 2,279 rows and 13 columns:
\describe{
\item{id}{Patient indentifier}
\item{pstate1, pstate2, pstate3, pstate4, pstate5, pstate6}{Predicted transition probabilities of transitions into states 1 to 6}
\item{se1, se2, se3, se4, se5, se6}{Standard error of the predicted transition probabilities of transitions into states 1 to 6}
}
}
}
\usage{
tp.cmprsk.j0
}
\description{
Used in vignette: Comparison-with-graphical-calibration-curves-in-competing-risks-setting. The predicted transition probabilities
are made at out of state j = 1 at time s = 0, and treat all states as absorbing states.
}
\keyword{datasets}
